requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
Pillow>=9.0.0
numpy>=1.21.0

# Configuration management
PyYAML>=6.0
pydantic>=2.0.0

# Server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
requests>=2.25.0

# Utilities
pygame>=2.1.0  # For examples
tqdm>=4.60.0
